+++ src/main/java/api/m2/file/record/FileDTO.java
@@ -21,7 +21,9 @@ public record FileDTO(
             LocalDateTime createdAt,
             FileType type,
             String contentType,
-            String checksum
+            String checksum,
+            boolean favorite,
+            LocalDateTime lastAccessedAt
     ) {
     }
 }