+++ src/main/java/api/m2/file/record/SetFavoriteRequest.java
@@ -1,0 +1,4 @@
+package api.m2.file.record;
+
+public record SetFavoriteRequest(boolean favorite) {
+}